refactor: renames directory to match component
Relates #37
